requirements.txt
Django==3.2.8
django-auth-ldap==2.2.0
django-crispy-forms==1.9.2
gunicorn==20.0.4
python-ldap==3.3.1
psycopg2-binary==2.8.6
djangorestframework==3.14.0
django-scheduler==0.10.1
schedule==1.2.0